Split endpoints for prediction
